Move form steps
